Add debian folder
Add a debian folder to build avocado-framework as
python3-avocado-framework package on Ubuntu.
Currently the debian/copyright file needs a review, among other things
to be improved before a final commit.

This commit is meant to be reviewed by upstream maintainers of
avocado-framework and work on improvements and pieces missing, so that
we can squash and do a proper commit at a later stage.

Co-authored-by: Arif Ali <arif.ali@canonical.com>
Signed-off-by: David Negreira <david.negreira@canonical.com>
Signed-off-by: Arif Ali <arif.ali@canonical.com>